Initial commit. Placeholder RoboFont observer code.
Start with RoboFont's sample code for creating an observer. The goal is
to observe key presses in Space Center and equalise sidebearings when a
special key is pressed so that this functionality can be triggered from
the keyboard instead of needing to open the control-click contextual
menu and selecting it from the menu options.