observer_test.py: Convert to new Python class
Inherit from `object` instead of not inheriting from anything to get a
new Python class. Actually, now that I'm writing this message, I'm
wondering if this may not be necessary since Python 3 defaults to new
Python objects regardless of how you write it. But then again this is
RoboFont and I'm not sure which Python version it's built against (was
thinking of TruFont for a moment since I'm currently trying to build
it).