observer_test.py: Center on 'a' key press
A lot of testing code from when I was trying to get this to work.
Actually just trying to inspect the objects/event.

Unfortunately, `print`ing and `pdb` don't seem to work. Fortunately
managed to get enough information from the `ObjectBrowser` once I
figured out that I could use that for inspecting.

The `print 'test'` at the end was just for me to confirm that the script
was running when I couldn't see any outward evidence of that fact.

Finally figured out what was going on and now the script will center
when you press 'a' in the Space Center.

Future enhancements include:
* Add undo support
* Make sure `center()` does what we're looking for and actually
  equalises sidebearings
* Add a setting that allows users of the plugin to customise the hotkey
  through a custom interface (and ideally through a text configuration
  file as well).