equalize_sidebearings.py: Copy RoboFont implementation
Copy `equalize_sidebearings` method and `getDefault` function from
RoboFont after uncompiling the `lib/UI/spaceCenter/*.pyc`,
`lib/doodleMenus.pyc`, and `lib/tools/defaults.pyc` files.

This gives us the exact same functionality that happens when choosing
"Equalize sidebearings" from the contextual menu in the Space Center.

Including the code directly instead of referencing it because the
`SpaceCenterMenuForGlyph` class opens a contextual menu on init, which
we really don't want, and the `equalizeSideBearings_` method that it
defined is an instance method, so I would need an instance of the class
but don't want to trigger the contextual menu. So I figured I'd just
copy the code directly to speed things up.

Unfortunately undo functionality doesn't work with the method I copied,
and I can't seem to figure out how to get it sorted out.