equalize_sidebearings.py: Get rid of proprietary code
Remove RoboFont's proprietary `getDefault` and `equalSideBearings_` code
since the extension now works without them and we shouldn't be using
this code in an extension that we plan to open source.

Sorry for even including this in the repo! It did end up being useful
for testing, just to see that it was possible.