build.sh: Update `timeStamp` on build
When building a new bundle, update the `timeStamp` key in `info.plist`
to correspond to the current time. Was hoping to retain the decimal
value in the timestamp, but for some reason only the integer part stays
after it gets set. Oh well, I don't care enough to fiddle with that.