preferences_window.py: Attempts to limit input character length
Tried to create an `NSFormatter` subclass but kept getting this stupid
error:

    TypeError: isPartialStringValid:newEditingString:errorDescription::
    Need tuple of 3 arguments as result

and:

    TypeError:
    isPartialStringValid:proposedSelectedRange:originalString:originalSelectedRange:errorDescription::
    Need tuple of 4 arguments as result

which is freaking idiotic because if you look at the Cocoa documentation
you can see that both those methods return fraking `BOOL`s.

Also needed to keep renaming the subclass because every time I
reinstalled the extension, PyObjC or RoboFont or whatever would yell:

    Traceback (most recent call last):
      File "preferences_window.py", line 10, in <module>
      error: Booya2HotKeyFormatter is overriding existing Objective-C
      class

Anyway, ended up ditching the `NSFormatter` subclass because of that
_stupid_ nonsensical error and am now just checking for string length in
the input edit callback and forcing a max length there by resetting the
value to ensure it doesn't go over 1 character.